refactor: move todos os service implementations para pasta service
